OOP_project/requirements.txt

google-generativeai
jupyter
python-dotenv
ipykernel
